requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-genai==4.2.2
langchain-text-splitters==1.1.2
google-genai==1.73.1
python-dotenv==1.0.1
pydantic>=2.0.0
requests>=2.32.5
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.12
langchain-chroma>=1.1.0
chromadb>=1.3.5
pypdf==4.3.1
streamlit==1.40.0